+++ wow_dbc/CHANGELOG.md
@@ -20,6 +20,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * ints in `Spell.dbc` to enums/flags.
 * ints in `Spell.dbc` for TBC/Wrath to `EffectAura`.
 * BREAKING: `EffectAura` renamed to `AuraMod`.
+* BREAKING: `BaseLanguage` removed and changed to `Language`.
+* BREAKING: `PowerType` renamed to `Power`.
 
 ### Removed